backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
psycopg[binary]==3.2.3
django-cors-headers==4.6.0
requests==2.32.3
